+++ frontend/src/App.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { RecoilRoot } from 'recoil';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
-import LoginMain from './components/LoginPage/LoginMain';
+import LoginMain from './components/Login/LoginMain';
 import CamMain from './components/Cam/CamMain';
 import BoostCamMain from './components/Main/BoostCamMain';
-import LoginCallback from './components/LoginPage/LoginCallback';
+import LoginCallback from './components/Login/LoginCallback';
 
 function App(): JSX.Element {
   return (
